[PureGo] Move example data helpers
