+++ purego/examples/dynamic/batch/main.go
@@ -1,6 +1,7 @@
 // Batch dynamic-proto ingestion example.
 // See examples/README.md for setup.
 // Run from the examples directory:
+//
 //	go run ./dynamic/batch
 package main
 
@@ -9,6 +10,7 @@ import (
 	"log"
 
 	"github.com/databricks/zerobus-sdk/purego/examples/config"
+	"github.com/databricks/zerobus-sdk/purego/examples/internal/exampleutil"
 	"github.com/databricks/zerobus-sdk/purego/zerobus"
 )
 
@@ -29,11 +31,11 @@ func main() {
 	}
 	defer stream.Close()
 
-	now := config.NowMicros()
+	now := exampleutil.NowMicros()
 	records := [][]byte{
-		[]byte(config.MakeOrderJSON(1, "Alice Smith", "Wireless Mouse", 2, 25.99, "pending", now)),
-		[]byte(config.MakeOrderJSON(2, "Bob Johnson", "Mechanical Keyboard", 1, 89.99, "shipped", now)),
-		[]byte(config.MakeOrderJSON(3, "Carol Williams", "USB-C Hub", 3, 45.00, "delivered", now)),
+		[]byte(exampleutil.MakeOrderJSON(1, "Alice Smith", "Wireless Mouse", 2, 25.99, "pending", now)),
+		[]byte(exampleutil.MakeOrderJSON(2, "Bob Johnson", "Mechanical Keyboard", 1, 89.99, "shipped", now)),
+		[]byte(exampleutil.MakeOrderJSON(3, "Carol Williams", "USB-C Hub", 3, 45.00, "delivered", now)),
 	}
 
 	batchOffset, err := stream.IngestJSONRecordsOffset(records)
