+++ purego/examples/dynamic/single/main.go
@@ -1,6 +1,7 @@
 // Single-record dynamic-proto ingestion example.
 // See examples/README.md for setup.
 // Run from the examples directory:
+//
 //	go run ./dynamic/single
 package main
 
@@ -9,6 +10,7 @@ import (
 	"log"
 
 	"github.com/databricks/zerobus-sdk/purego/examples/config"
+	"github.com/databricks/zerobus-sdk/purego/examples/internal/exampleutil"
 	"github.com/databricks/zerobus-sdk/purego/zerobus"
 )
 
@@ -33,11 +35,11 @@ func main() {
 	}
 	defer stream.Close()
 
-	now := config.NowMicros()
+	now := exampleutil.NowMicros()
 	records := [][]byte{
-		[]byte(config.MakeOrderJSON(1, "Alice Smith", "Wireless Mouse", 2, 25.99, "pending", now)),
-		[]byte(config.MakeOrderJSON(2, "Bob Johnson", "Mechanical Keyboard", 1, 89.99, "shipped", now)),
-		[]byte(config.MakeOrderJSON(3, "Carol Williams", "USB-C Hub", 3, 45.00, "delivered", now)),
+		[]byte(exampleutil.MakeOrderJSON(1, "Alice Smith", "Wireless Mouse", 2, 25.99, "pending", now)),
+		[]byte(exampleutil.MakeOrderJSON(2, "Bob Johnson", "Mechanical Keyboard", 1, 89.99, "shipped", now)),
+		[]byte(exampleutil.MakeOrderJSON(3, "Carol Williams", "USB-C Hub", 3, 45.00, "delivered", now)),
 	}
 	for i, rec := range records {
 		offset, err := stream.IngestJSONOffset(rec)
