+++ purego/examples/json/batch/main.go
@@ -22,6 +22,7 @@ import (
 	"sync/atomic"
 
 	"github.com/databricks/zerobus-sdk/purego/examples/config"
+	"github.com/databricks/zerobus-sdk/purego/examples/internal/exampleutil"
 	"github.com/databricks/zerobus-sdk/purego/zerobus"
 )
 
@@ -56,13 +57,13 @@ func main() {
 	}
 	defer stream.Close()
 
-	now := config.NowMicros()
+	now := exampleutil.NowMicros()
 
 	// Build and queue one batch.
 	batch := [][]byte{
-		[]byte(config.MakeOrderJSON(1, "Alice Smith", "Wireless Mouse", 2, 25.99, "pending", now)),
-		[]byte(config.MakeOrderJSON(2, "Bob Johnson", "Mechanical Keyboard", 1, 89.99, "shipped", now)),
-		[]byte(config.MakeOrderJSON(3, "Carol Williams", "USB-C Hub", 3, 45.00, "delivered", now)),
+		[]byte(exampleutil.MakeOrderJSON(1, "Alice Smith", "Wireless Mouse", 2, 25.99, "pending", now)),
+		[]byte(exampleutil.MakeOrderJSON(2, "Bob Johnson", "Mechanical Keyboard", 1, 89.99, "shipped", now)),
+		[]byte(exampleutil.MakeOrderJSON(3, "Carol Williams", "USB-C Hub", 3, 45.00, "delivered", now)),
 	}
 	batchOffset, err := stream.IngestRecordsOffset(batch)
 	if err != nil {
