+++ purego/examples/proto/batch/main.go
@@ -22,6 +22,7 @@ import (
 	"google.golang.org/protobuf/proto"
 
 	"github.com/databricks/zerobus-sdk/purego/examples/config"
+	"github.com/databricks/zerobus-sdk/purego/examples/internal/exampleutil"
 	"github.com/databricks/zerobus-sdk/purego/examples/proto/pb"
 	"github.com/databricks/zerobus-sdk/purego/zerobus"
 )
@@ -61,7 +62,7 @@ func main() {
 	}
 	defer stream.Close()
 
-	now := config.NowMicros()
+	now := exampleutil.NowMicros()
 
 	// Marshal records and queue one batch.
 	orders := []*pb.Order{
