+++ purego/examples/proto/runtime/main.go
@@ -16,6 +16,7 @@ import (
 	"google.golang.org/protobuf/types/dynamicpb"
 
 	"github.com/databricks/zerobus-sdk/purego/examples/config"
+	"github.com/databricks/zerobus-sdk/purego/examples/internal/exampleutil"
 	"github.com/databricks/zerobus-sdk/purego/zerobus"
 )
 
@@ -118,7 +119,7 @@ func main() {
 	}
 	defer stream.Close()
 
-	now := config.NowMicros()
+	now := exampleutil.NowMicros()
 	orders := []struct {
 		id       int32
 		customer string
